requirements.txt
streamlit>=1.20.0
openai>=0.27.0
pinecone-client>=2.2.0
whisper-openai>=20230214.1
deepseek-api>=1.0.0
google-api-core>=2.11.0
oauthlib>=3.2.0
requests>=2.28.0
tenacity>=8.1.0
python-dotenv>=0.21.0
pandas>=1.5.0
numpy>=1.24.0